better single-chars and KeyChords
Both `atom` and `atom_i` (inverted version of `atom`) are redundant,
because only the inverted one was used.  Now there's just the one.

Chord handling improved to the point where I caught a bug (no `C-v` to
select visual-block-mode) and adding the functionality into the existing
code was a breeze.